+++ lib/Controller/ExAppsPageController.php
@@ -10,6 +10,7 @@ use OC\App\DependencyAnalyzer;
 use OC\App\Platform;
 use OC_App;
 use OCA\AppAPI\AppInfo\Application;
+use OCA\AppAPI\Attribute\AppAPIAuth;
 use OCA\AppAPI\Db\DaemonConfig;
 use OCA\AppAPI\Db\ExApp;
 use OCA\AppAPI\Db\ExAppScope;
@@ -23,8 +24,10 @@ use OCA\AppAPI\Service\ExAppUsersService;
 use OCP\App\IAppManager;
 use OCP\AppFramework\Controller;
 use OCP\AppFramework\Http;
+use OCP\AppFramework\Http\Attribute\NoAdminRequired;
 use OCP\AppFramework\Http\Attribute\NoCSRFRequired;
 use OCP\AppFramework\Http\Attribute\PasswordConfirmationRequired;
+use OCP\AppFramework\Http\Attribute\PublicPage;
 use OCP\AppFramework\Http\ContentSecurityPolicy;
 use OCP\AppFramework\Http\JSONResponse;
 use OCP\AppFramework\Http\TemplateResponse;
@@ -101,13 +104,16 @@ class ExAppsPageController extends Controller {
 		$appInitialData = [
 			'appstoreEnabled' => $this->config->getSystemValueBool('appstoreenabled', true),
 			'updateCount' => count($this->getExAppsWithUpdates()),
-			'default_daemon_config' => $defaultDaemonConfigName,
-			'docker_socket_accessible' => $this->dockerActions->isDockerSocketAvailable(),
 		];
 
 		if ($defaultDaemonConfigName !== '') {
 			$daemonConfig = $this->daemonConfigService->getDaemonConfigByName($defaultDaemonConfigName);
-			$appInitialData['daemon_config'] = $daemonConfig;
+			$this->dockerActions->initGuzzleClient($daemonConfig);
+			$daemonConfigAccessible = $this->dockerActions->ping($this->dockerActions->buildDockerUrl($daemonConfig));
+			$appInitialData['daemon_config_accessible'] = $daemonConfigAccessible;
+			if (!$daemonConfigAccessible) {
+				$this->logger->error(sprintf('Deploy daemon "%s" is not accessible by Nextcloud. Please verify its configuration', $daemonConfig->getName()));
+			}
 		}
 
 		$this->initialStateService->provideInitialState('apps', $appInitialData);
@@ -256,6 +262,7 @@ class ExAppsPageController extends Controller {
 				'daemon' => $daemon,
 				'systemApp' => $exApp !== null && $this->exAppUsersService->exAppUserExists($exApp->getAppid(), ''),
 				'exAppUrl' => $exApp !== null ? AppAPIService::getExAppUrl($exApp->getProtocol(), $exApp->getHost(), $exApp->getPort()) : '',
+				'status' => $exApp !== null ? json_decode($exApp->getStatus(), true) : [],
 			];
 		}
 
@@ -398,6 +405,7 @@ class ExAppsPageController extends Controller {
 					'exAppUrl' => AppAPIService::getExAppUrl($exApp->getProtocol(), $exApp->getHost(), $exApp->getPort()),
 					'releases' => [],
 					'update' => null,
+					'status' => json_decode($exApp->getStatus(), true),
 				];
 			}
 		}
@@ -447,19 +455,25 @@ class ExAppsPageController extends Controller {
 						// 2. Register ExApp (container must be already initialized successfully)
 						if (!$this->registerExApp($appId, $infoXml, $daemonConfig)) {
 							$this->service->unregisterExApp($appId); // Fallback unregister if failure
+							return new JSONResponse(['data' => ['message' => $this->l10n->t('Failed to register ExApp')]], Http::STATUS_INTERNAL_SERVER_ERROR);
 						}
 					} else {
-						$this->logger->error('Failed to Deploy ExApp');
+						$this->logger->error(sprintf('Failed to Deploy %s ExApp', $appId));
 						return new JSONResponse([
 							'data' => [
-								'message' => 'Failed to Deploy ExApp',
+								'message' => $this->l10n->t('Failed to Deploy ExApp'),
 							]
 						], Http::STATUS_INTERNAL_SERVER_ERROR);
 					}
 
 					$exApp = $this->service->getExApp($appId);
+
+					// Start ExApp initialization step (to download dynamic content, e.g. models)
+					$this->service->dispatchExAppInit($exApp);
+
+					// TODO: Return intermediate status for UI, so that it knows when to call separate enable endpoint to finish enable action
 					if (!$this->service->enableExApp($exApp)) {
-						return new JSONResponse(['data' => ['message' => 'Failed to enable ExApp']], Http::STATUS_INTERNAL_SERVER_ERROR);
+						return new JSONResponse(['data' => ['message' => $this->l10n->t('Failed to enable ExApp')]], Http::STATUS_INTERNAL_SERVER_ERROR);
 					}
 					return new JSONResponse([
 						'data' => [
@@ -481,7 +495,7 @@ class ExAppsPageController extends Controller {
 				}
 
 				if (!$this->service->enableExApp($exApp)) {
-					return new JSONResponse(['data' => ['message' => 'Failed to enable ExApp']], Http::STATUS_INTERNAL_SERVER_ERROR);
+					return new JSONResponse(['data' => ['message' => $this->l10n->t('Failed to enable ExApp')]], Http::STATUS_INTERNAL_SERVER_ERROR);
 				}
 			}
 
@@ -587,7 +601,7 @@ class ExAppsPageController extends Controller {
 			foreach ($appIds as $appId) {
 				$exApp = $this->service->getExApp($appId);
 				if (!$this->service->disableExApp($exApp)) {
-					return new JSONResponse(['data' => ['message' => 'Failed to disable ExApp']], Http::STATUS_INTERNAL_SERVER_ERROR);
+					return new JSONResponse(['data' => ['message' => $this->l10n->t('Failed to disable ExApp')]], Http::STATUS_INTERNAL_SERVER_ERROR);
 				}
 			}
 			return new JSONResponse([]);
@@ -650,15 +664,40 @@ class ExAppsPageController extends Controller {
 			'host' => $exAppInfo['host'],
 			'port' => (int) $exAppInfo['port'],
 		])) {
+			// TODO: Set ExApp status before initialization to display progress in UI
+			// 5.5 Dispatch init step on ExApp side
+			$this->service->dispatchExAppInit($exApp);
+
+			return new JSONResponse([
+				'data' => [
+					'appid' => $appId,
+				]
+			]);
+
 			// 6. Enable ExApp
-			$this->service->enableExApp($exApp);
+			//$this->service->enableExApp($exApp);
+		}
+
+		//return new JSONResponse([
+		//	'data' => [
+		//		'appid' => $appId,
+		//		'systemApp' => filter_var($exAppInfo['system_app'], FILTER_VALIDATE_BOOLEAN),
+		//		'exAppUrl' => AppAPIService::getExAppUrl($exAppInfo['protocol'], $exAppInfo['host'], (int) $exAppInfo['port']),
+		//	]
+		//]);
+	}
+
+	public function enableExApp(string $appId): JSONResponse {
+		$exApp = $this->service->getExApp($appId);
+		if (!$this->service->enableExApp($exApp)) {
+			return new JSONResponse(['data' => ['message' => $this->l10n->t('Failed to enable ExApp')]]);
 		}
 
 		return new JSONResponse([
 			'data' => [
 				'appid' => $appId,
-				'systemApp' => filter_var($exAppInfo['system_app'], FILTER_VALIDATE_BOOLEAN),
-				'exAppUrl' => AppAPIService::getExAppUrl($exAppInfo['protocol'], $exAppInfo['host'], (int) $exAppInfo['port']),
+				'systemApp' => $this->exAppUsersService->exAppUserExists($exApp->getAppid(), ''),
+				'exAppUrl' => AppAPIService::getExAppUrl($exApp->getProtocol(), $exApp->getHost(), $exApp->getPort()),
 			]
 		]);
 	}
@@ -727,6 +766,16 @@ class ExAppsPageController extends Controller {
 	 */
 	public function listCategories(): JSONResponse {
 		return new JSONResponse($this->getAllCategories());
+	}
+
+	/**
+	 * Get ExApp status, that includes initialization information
+	 *
+	 * @return JSONResponse
+	 */
+	public function getAppStatus(string $appId): JSONResponse {
+		$exApp = $this->service->getExApp($appId);
+		return new JSONResponse(json_decode($exApp->getStatus(), true));
 	}
 
 	/**