+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [1.1.0 - 2023-10-20]
+
+### Added
+
+- Added ExApp initialization progress
+- Added disabled state of app management actions if Deploy daemon is not accessible
+- Added support for new fileActions registration (Nextcloud 28)
+
 ## [1.0.1 - 2023-10-06]
 
 ### Fixed